requirements.txt
opencv-python
pyautogui
numpy
opencv-python
Pillow
pygame
